refactor(.github/scorecard): trim rubric samples and extract shared lib
