--- .github/scorecard/.gitignore
@@ -1,1 +1,0 @@
-.scorecard-results.json
